+++ lib/providers/firebase_cache_provider.g.dart
@@ -121,4 +121,4 @@ final class FirebaseCacheProvider
   }
 }
 
-String _$firebaseCacheHash() => r'68e7d990c003635c751a762a59127cc753c8bade';
+String _$firebaseCacheHash() => r'5f67df1a526aff39b5a122585b46608e6609b098';